Refactor string formatting in ReplSession
